Add LICENSE.txt with inherited static linking exception
As per the LGPL laz-rs is considered a modification of the LASzip library, thus laz-rs must be licensed under the same terms as LASzip.

LGPL only really defines how we are allowed to use the library with dynamic linking (thus static linking is not allowed).

rustc (the Rust compiler) only does static linking (between pure Rust libs).
LASzip has an exception in its LICENSE.txt allowing static linking and also allowing to extent this exception to our modification of the library.

We include that exception in the LICENSE because we basically need it.

I'm not lawyer but that's what I understand :thinking:

https://github.com/LAStools/LAStools/issues/6
https://groups.google.com/forum/#!topic/lastools/Bs9YBXzXXcg
https://github.com/LAStools/LAStools/commit/76bdb0eff89912a40b10743c0ce7853ceeafb312