Improve ParLasZipDecompressor performances
Change the implementation of `ParLasZipDecompressor` to not use
the "outernal" buffer anymore.

The "outernal" buffer was used to decompress points into it, then
copy the decompressed data into the caller's buffer.

This meant that when the caller wanted to decompress a lot of data
like decompress a 3GB LAZ to 16GB LAS (by passing a slice of 16G
to decompress_many), the "outernal" buffer used was also 16G.
Which meant that 32GB were allocated, and the decompress_many would
also need to copy 16GB of data from the outernal buffer to the callers
buffer. yikes

This commit removes the use of this internal buffer so the result of the
decompression is, as much as possible, directly written into
 the caller's buffer.

This saves memory usage, and reduces copy overhead.