Fix seeking to point in last chunk for fmt >= 6
There is a safeguard in place to catch a potential error when seeking to
a point that belongs to the last chunk.

This safeguard basically checks that after decompressing a point
we are not at the beginning or after the chunk table (meaning
the point we want to seek to does not exist)

Point format >= 6 uses Layered chunks, where on first decompression of a
point, each layer reads all the necessary data into memory.
Meaning that after decompressing the first point of the last
chunk, our checks believe we are at the beginning of the chunk table
thus it believes the point does not exist.

To fix this, we simply don't do this safeguard check for layered chunked
points. If the point does not exist, a normal eof error will be returned
as the in memory buffer of the layer would have been fully consumed.