Fix default value for point6 changed flags
The compression of each field in point format >= 6
happens in its own byte buffer, and that buffer is only written
to the file if the value of the field changed and was not always the
same.

The flags used to track whether a field changed were initialized to
`true`, meaning that even if the field values were always the same,
the buffer corresponding to the field would get written in the file.

This was mostly ok, as most of the fields generally have some
change in their values, or, all values were zero, which would still
act 'nice'

However, if a field had all its values the same and non zero, this would
lead to incorrect compression result. (e.g a file with
point_source_id.all() == 1).

The fix is to correctly initialize to `false`,
This only affected point format >= 6 && <= 10, and only the fields from
the point format 6 (since point format >= 7 are based on format 6).