Change AC_BUFFER_SIZE to 4096
This is the value used in the original LASzip.

And with a value of 1024 like we had, the result of the encoding
of some fields would not be the same than LASZip.

Note that if we change LASZip's AC_BUFFER_SIZE to 1024, the result
of the encoding would become the same as what we get with
AC_BUFFER_SIZE=1024.

Seems like the field mostly affected (potentially) is user_data