Correctly use has_changed tracking
Layered compressors tracks if data actually changed
as even if the data did not change, compressors will output some bytes

We were not correctly using that information when writing layer sizes
and layer bytes, leading to us writing more bytes than laszip would.

This however did not harm in any way as it could only happen on layer
which did not change, and as we wrote a size which was correct with
how many bytes we wrote, thus, other impl (and us) did not trip on it.

This change is 'cosmetic', just to have this impl write the exact same
number of bytes than laszip would.