Make the LazAppender require the point count
The point count is needed as some files maybe have extra data in the
chunks which we may decompress as points while its not.

Leading to appending not being completely correct,
so now the LazAppender requires to get the number of points in the file.

Actually this number of points is where the appender should be ready to
append, so it could be used to "truncate" , then append