ParLasZipDecompressor: use read_exact not read
We were using `read` instead of `read_exact` when reading from
the source to fill our internal buffer.

However read does not guarantee that it will fill the output buffer, and
intead returns how many bytes it read.

Our intent was to use read_exact, which is now the case.

The bug when triggered, meant that the internal buffer was incomplete
leading to decompressor decompressing wrong data and creating EOF errors.

Luckily to trigger the bug some the source must be in non-standard conditions.
For example, as often the source is a BufReader, but with a buffer capacity
>= chunk size, which is not the defaul setting. This would lead to
the BufReader empyting its buffer, but not making sur our buffer was
filled.